+++ database/migrations/2023_04_02_035103_create_viatico_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->unsignedBigInteger('user_id')->nullable(true);
             $table->foreign('user_id')->references('id')->on('users');
             $table->date('fecha_viatico');
-            $table->float('total_viatico')->default(0)->nullable(true);
+            $table->decimal('total_viatico',10,2)->default(0)->nullable(true);
             $table->string('observaciones')->nullable(true);
             $table->timestamps();
         });
